urllib.parse.urljoin messes it up entirely if using http
<soren> RickXy: First, it rewrites git@github.com:JioCloud/whatever.git
to gopher://git@github.com:JioCloud/whatever.git (it adds the gopher://
thing to make urllib.parse happy).
<soren> RickXy: Than it does urljoin with the url (i.e. "..").
<soren> RickXy: This assumes that whatever.git is a filename (since
there's no / at the end (and if there had been, this method would have
stripped it off).
<soren> RickXy: So ".." brings us to... the root directory.
<soren> Not JioCloud.
<soren> This is done correctly for https, but urlparse sees the
"JioCloud" part as a port(!), so it's not part of the URL that gets
joined with "..".
<soren> ...and since it's a port, you can't strip it off with "..".
<RickXy> :-)
<soren> RickXy: Make it "." instead of ".." and it works for both.